# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Books Library Management System",
        "description": "API for managing a library of books and reviews, with recommendations based on the genre and average rating of a provided book title.",
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "summary": "Root",
                "description": "Welcome endpoint.\n\nReturns:\n    dict: A welcome message.",
                "operationId": "root__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/books": {
            "get": {
                "summary": "Get Books",
                "description": "Get a list of all books in the library.\n\nReturns:\n    List[Book]: A list of all books.",
                "operationId": "get_books_books_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    },
                                    "type": "array",
                                    "title": "Response Get Books Books Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Add Book",
                "description": "Add a new book to the library.\n\nArgs:\n    book (Book): The book details.\n    username (str): The username of the authenticated user.\n\nReturns:\n    dict: The added book details including the ID.",
                "operationId": "add_book_books_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBasic": []
                    }
                ]
            }
        },
        "/books/{id}": {
            "get": {
                "summary": "Get Book",
                "description": "Get a book by its ID.\n\nArgs:\n    id (str): The ID of the book as a valid MongoDB ObjectId.\n\nReturns:\n    Book: The book details.\n\nRaises:\n    HTTPException: If the book is not found or the ID format is invalid.",
                "operationId": "get_book_books__id__get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the book as a valid MongoDB ObjectId",
                            "title": "Id"
                        },
                        "description": "The ID of the book as a valid MongoDB ObjectId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Book",
                "description": "Update a book's details by its ID.\n\nArgs:\n    id (str): The ID of the book as a valid MongoDB ObjectId.\n    book (Book): The updated book details.\n    username (str): The username of the authenticated user.\n\nReturns:\n    Book: The updated book details.\n\nRaises:\n    HTTPException: If the book is not found or the ID format is invalid.",
                "operationId": "update_book_books__id__put",
                "security": [
                    {
                        "HTTPBasic": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the book as a valid MongoDB ObjectId",
                            "title": "Id"
                        },
                        "description": "The ID of the book as a valid MongoDB ObjectId"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Book",
                "description": "Delete a book by its ID.\n\nArgs:\n    id (str): The ID of the book as a valid MongoDB ObjectId.\n    username (str): The username of the authenticated user.\n\nReturns:\n    dict: A message indicating the book was deleted.\n\nRaises:\n    HTTPException: If the book is not found or the ID format is invalid.",
                "operationId": "delete_book_books__id__delete",
                "security": [
                    {
                        "HTTPBasic": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the book as a valid MongoDB ObjectId",
                            "title": "Id"
                        },
                        "description": "The ID of the book as a valid MongoDB ObjectId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/reviews/{book_id}": {
            "post": {
                "summary": "Add Review",
                "description": "Add a review to a book.\n\nArgs:\n    book_id (str): The ID of the book as a valid MongoDB ObjectId.\n    review (Review): The review details.\n    username (str): The username of the authenticated user.\n\nReturns:\n    dict: The added review details including the ID.\n\nRaises:\n    HTTPException: If the book is not found or the ID format is invalid.",
                "operationId": "add_review_reviews__book_id__post",
                "security": [
                    {
                        "HTTPBasic": []
                    }
                ],
                "parameters": [
                    {
                        "name": "book_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the book as a valid MongoDB ObjectId",
                            "title": "Book Id"
                        },
                        "description": "The ID of the book as a valid MongoDB ObjectId"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Review"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Review"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Get Reviews",
                "description": "Get a list of reviews for a specific book.\n\nArgs:\n    book_id (str): The ID of the book as a valid MongoDB ObjectId.\n\nReturns:\n    List[Review]: A list of reviews for the book.\n\nRaises:\n    HTTPException: If the ID format is invalid.",
                "operationId": "get_reviews_reviews__book_id__get",
                "parameters": [
                    {
                        "name": "book_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the book as a valid MongoDB ObjectId",
                            "title": "Book Id"
                        },
                        "description": "The ID of the book as a valid MongoDB ObjectId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Review"
                                    },
                                    "title": "Response Get Reviews Reviews  Book Id  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recommendations/{book_title}": {
            "get": {
                "summary": "Get Recommendations",
                "description": "Get book recommendations based on the genre and average rating of a provided book title.\n\nArgs:\n    book_title (str): The title of the book.\n\nReturns:\n    list: A list of recommended books.",
                "operationId": "get_recommendations_recommendations__book_title__get",
                "parameters": [
                    {
                        "name": "book_title",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The title of the book for generating recommendations based on genre and average rating",
                            "title": "Book Title"
                        },
                        "description": "The title of the book for generating recommendations based on genre and average rating"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Book": {
                "properties": {
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "author": {
                        "type": "string",
                        "title": "Author"
                    },
                    "genre": {
                        "type": "string",
                        "title": "Genre"
                    },
                    "year_published": {
                        "type": "integer",
                        "title": "Year Published"
                    },
                    "summary": {
                        "type": "string",
                        "title": "Summary"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "author",
                    "genre",
                    "year_published",
                    "summary"
                ],
                "title": "Book"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Review": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "title": "User Id"
                    },
                    "review_text": {
                        "type": "string",
                        "title": "Review Text"
                    },
                    "rating": {
                        "type": "number",
                        "title": "Rating"
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "review_text",
                    "rating"
                ],
                "title": "Review"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBasic": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}